refactor: 기존 UI 컴포넌트들 features 레이어까지 분할
