+++ src/refactoring/features/cart/ui/CartList.tsx
@@ -1,0 +1,27 @@
+import { CartItem } from './CartItem.tsx';
+import { ICartItem } from '../../../app/types';
+
+interface CartListProps {
+  cart: ICartItem[];
+  updateQuantity: (productId: string, quantity: number) => void;
+  removeFromCart: (productId: string) => void;
+}
+
+export function CartList({
+  cart,
+  updateQuantity,
+  removeFromCart,
+}: CartListProps) {
+  return (
+    <div className="space-y-2">
+      {cart.map((item) => (
+        <CartItem
+          key={item.product.id}
+          item={item}
+          updateQuantity={updateQuantity}
+          removeFromCart={removeFromCart}
+        />
+      ))}
+    </div>
+  );
+}
