+++ src/refactoring/widgets/product/ui/ProductManagement.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react';
 import { SectionTitle } from '../../../shared/ui/typography';
 import { Input } from '../../../shared/ui/inputs';
 import { IDiscount, IProduct } from '../../../app/types';
+import { ProductForm } from '../../../features/product/ui/ProductForm.tsx';
 
 interface ProductManagementProps {
   products: IProduct[];
@@ -20,25 +21,6 @@ export function ProductManagement({
     quantity: 0,
     rate: 0,
   });
-  const [newProduct, setNewProduct] = useState<Omit<IProduct, 'id'>>({
-    name: '',
-    price: 0,
-    stock: 0,
-    discounts: [],
-  });
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-
-  const handleAddNewProduct = () => {
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-    onProductAdd(productWithId);
-    setNewProduct({
-      name: '',
-      price: 0,
-      stock: 0,
-      discounts: [],
-    });
-    setShowNewProductForm(false);
-  };
 
   const toggleProductAccordion = (productId: string) => {
     setOpenProductIds((prev) => {
@@ -118,77 +100,7 @@ export function ProductManagement({
   return (
     <div>
       <SectionTitle title={'상품 관리'} />
-      <button
-        onClick={() => setShowNewProductForm(!showNewProductForm)}
-        className="bg-green-500 text-white px-4 py-2 rounded mb-4 hover:bg-green-600"
-      >
-        {showNewProductForm ? '취소' : '새 상품 추가'}
-      </button>
-      {showNewProductForm && (
-        <div className="bg-white p-4 rounded shadow mb-4">
-          <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
-          <div className="mb-2">
-            <label
-              htmlFor="productName"
-              className="block text-sm font-medium text-gray-700"
-            >
-              상품명
-            </label>
-            <Input
-              id={'productName'}
-              type={'text'}
-              value={newProduct.name}
-              onChange={(e) =>
-                setNewProduct({ ...newProduct, name: e.target.value })
-              }
-            />
-          </div>
-          <div className="mb-2">
-            <label
-              htmlFor="productPrice"
-              className="block text-sm font-medium text-gray-700"
-            >
-              가격
-            </label>
-            <Input
-              id={'productPrice'}
-              type={'number'}
-              value={newProduct.price}
-              onChange={(e) =>
-                setNewProduct({
-                  ...newProduct,
-                  price: parseInt(e.target.value),
-                })
-              }
-            />
-          </div>
-          <div className="mb-2">
-            <label
-              htmlFor="productStock"
-              className="block text-sm font-medium text-gray-700"
-            >
-              재고
-            </label>
-            <Input
-              id={'productStock'}
-              type={'number'}
-              value={newProduct.stock}
-              onChange={(e) =>
-                setNewProduct({
-                  ...newProduct,
-                  stock: parseInt(e.target.value),
-                })
-              }
-            />
-          </div>
-          <button
-            onClick={handleAddNewProduct}
-            className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-          >
-            추가
-          </button>
-        </div>
-      )}
+      <ProductForm onProductAdd={onProductAdd} />
       <div className="space-y-2">
         {products.map((product, index) => (
           <div
